+++ packages/airgap.js-types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @transcend-io/airgap.js-types
 
+## 14.2.28
+
+### Patch Changes
+
+- Updated dependencies [188ba6f]
+  - @transcend-io/privacy-types@5.12.0
+
 ## 14.2.27
 
 ### Patch Changes